+++ avakas/__main__.py
@@ -1,0 +1,7 @@
+"""
+avakas main cli exec
+"""
+from avakas.cli import main
+
+if __name__ == "__main__":
+    main()
