+++ avakas/errors.py
@@ -1,0 +1,9 @@
+"""
+Avakas Errors
+"""
+
+
+class AvakasError(Exception):
+    """
+    Basic Avakas Error
+    """
