+++ setup.py
@@ -1,11 +1,14 @@
 #!/usr/bin/env python
-import sys, os
+import sys
+import os
+
 
 try:
     from setuptools import setup
 except ImportError:
     print("avakas requires setuptools")
     sys.exit(1)
+
 
 def main():
     vsn_path = "%s/version" % os.path.dirname(os.path.abspath(__file__))
@@ -25,10 +28,15 @@ def main():
           license='MIT',
           url='https://github.com/otakup0pe/avakas',
           install_requires=['semantic_version', 'gitpython', 'erl_terms'],
-          scripts=['avakas'],
+          packages=['avakas', 'avakas.flavors'],
+          entry_points={
+            'console_scripts': ['avakas = avakas.cli:main']
+          },
           include_package_data=True,
-          package_data={'avakas':['version']}
-         )
+          package_data={'avakas': [
+            'version'
+          ]})
+
 
 if __name__ == "__main__":
     main()
